More WorldServer packets! \ o /

+++ SharpFly Utility Library/Enums/Gender.cs
@@ -1,6 +1,6 @@
 ﻿namespace SharpFly_Utility_Library.Enums
 {
-    public enum Gender : uint
+    public enum Gender : byte
     {
         MALE = 0,
         FEMALE,
